fix(disable-smart-testing): Refactoring changes.
