+++ surefire-provider/src/main/java/org/arquillian/smart/testing/surefire/provider/Validate.java
@@ -9,12 +9,4 @@ public class Validate {
     public static boolean isNotEmpty(String string){
         return !isEmpty(string);
     }
-
-    public static boolean isDisableSmartTesting() {
-        String disableSmartTesting = System.getProperty("disableSmartTesting");
-        if (disableSmartTesting != null && disableSmartTesting.equals("true")){
-            return true;
-        }
-        return false;
-    }
 }
